Make KVC syntax more robust
Allow `foo.%(...)` so that keys can be supplied via objects or as the result of a function or method call. Additionally, replacement is done with ranges of text instead of regular expressions based on what to replace. This is much less error-prone.

Also fixes some critical bugs that were revealed by these additions.