Support readonly %properties, use class_addProperty
With these additions, marking a %property as `readonly` will not synthesize a getter and setter.

Clang will not synthesize a backing ivar when you implement both the getter and setter of a property (or when you implement the getter of a readonly property) because it reasonably assumes you don't _need_ backing storage

Likewise, if you mark a %property as `readonly` Logos will assume that you plan to add a computed property which doesn't need backing storage. This makes sense since Logos's generated getters use associated objects, and you would need to provide some kind of setter-logic if you did need backing storage.